backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==0.8.11
pdfminer.six==20231228
scikit-learn==1.5.2
numpy==1.26.4
pydantic==2.7.4
nltk==3.9.1
transformers==4.43.0
torch==2.2.0

# Pin cryptography to an older CFFI-based build to avoid Windows DLL issues
cryptography==36.0.2
